Add Close Editor button, move Download GPX button
